refactor: inline random navigation providers
