+++ lib/features/scenes/presentation/pages/scene_details_page.dart
@@ -24,9 +24,8 @@ import '../../../../core/utils/responsive.dart';
 import '../../domain/entities/scene_title_utils.dart';
 import '../providers/entity_media_filter_scope.dart';
 import '../providers/scene_details_provider.dart';
-import '../providers/scene_list_provider.dart';
 import '../providers/playback_queue_provider.dart';
-import '../providers/scene_random_navigation_provider.dart';
+import '../providers/scene_list_provider.dart';
 import '../providers/video_player_provider.dart';
 import 'scene_info_page.dart';
 import '../../data/repositories/stream_resolver.dart';
@@ -105,7 +104,7 @@ class _SceneDetailsPageState extends ConsumerState<SceneDetailsPage> {
 
   Future<void> _openRandomScene(BuildContext context) async {
     final randomScene = await ref
-        .read(sceneRandomNavigationControllerProvider)
+        .read(sceneListProvider.notifier)
         .getRandomScene(excludeSceneId: widget.sceneId);
     if (!context.mounted) return;
