+++ lib/features/scenes/presentation/widgets/scene_video_player.dart
@@ -8,7 +8,7 @@ import 'package:go_router/go_router.dart';
 
 import '../../domain/entities/scene.dart';
 import '../../domain/entities/scene_title_utils.dart';
-import '../providers/scene_random_navigation_provider.dart';
+import '../providers/scene_list_provider.dart';
 import '../providers/player_view_mode.dart';
 import '../providers/player_settings.dart';
 import '../providers/video_player_provider.dart';
@@ -421,7 +421,7 @@ class _SceneVideoPlayerState extends ConsumerState<SceneVideoPlayer> {
 
   Future<void> _openRandomScene() async {
     final randomScene = await ref
-        .read(sceneRandomNavigationControllerProvider)
+        .read(sceneListProvider.notifier)
         .getRandomScene(excludeSceneId: widget.scene.id);
     if (!mounted || randomScene == null) return;
