+++ lib/features/tags/presentation/pages/tag_details_page.dart
@@ -16,7 +16,6 @@ import '../../../../core/presentation/theme/app_theme.dart';
 import '../../../setup/presentation/providers/navigation_customization_provider.dart';
 
 import '../providers/tag_list_provider.dart';
-import '../providers/tag_random_navigation_provider.dart';
 import 'package:stash_app_flutter/features/scenes/presentation/providers/playback_queue_provider.dart';
 import 'package:stash_app_flutter/features/scenes/presentation/widgets/scene_strip.dart';
 import 'package:stash_app_flutter/features/galleries/presentation/widgets/gallery_strip.dart';
@@ -44,7 +43,7 @@ class TagDetailsPage extends ConsumerWidget {
 
   Future<void> _openRandomTag(BuildContext context, WidgetRef ref) async {
     final randomTag = await ref
-        .read(tagRandomNavigationControllerProvider)
+        .read(tagListProvider.notifier)
         .getRandomTag(excludeTagId: tagId);
     if (!context.mounted) return;
